+++ src/runtime/manifest.ts
@@ -148,10 +148,12 @@ export interface ManifestStack {
   requiredAPIs: ManifestRequiredAPI[];
   endpoints: Record<string, ManifestEndpoint>;
   extensions?: Record<string, ManifestExtension>;
+  requiredRoles?: string[];
 }
 
 /**
  * Returns the JSON representation of a ManifestStack, which has CEL
+
  * expressions in its options as object types, with its expressions
  * transformed into the actual CEL strings.
  *